Crate printer

Re-exports

pub use pretty::termcolor;

Modules

ast 🔒

de_bruijn 🔒

latex

print_to_string 🔒

theme 🔒

tokens 🔒

types 🔒

Structs

ColorSpec

A color specification.

PrintCfg

StandardStream

Satisfies io::Write and WriteColor, and supports optional coloring to either of the standard output streams, stdout and stderr.

Enums

Color

The set of available colors for the terminal foreground/background.

ColorChoice

ColorChoice represents the color preferences of an end user.

Constants

DEFAULT_WIDTH

Traits

Print

PrintExt

PrintInCtx

PrintToString

PrintToStringInCtx

WriteColor

This trait describes the behavior of writers that support colored output.

Type Definitions

Alloc

Builder