Crate printer

Re-exports

pub use pretty::termcolor;

Modules

ast 🔒

ctx 🔒

de_bruijn 🔒

dec 🔒

env 🔒

fragments

latex

nf 🔒

print_to_string 🔒

theme 🔒

tokens 🔒

types 🔒

util 🔒

val 🔒

Structs

ColorSpec

A color specification.

PrintCfg

StandardStream

Satisfies io::Write and WriteColor, and supports optional coloring to either of the standard output streams, stdout and stderr.

Enums

Color

The set of available colors for the terminal foreground/background.

ColorChoice

ColorChoice represents the color preferences of an end user.

Constants

DEFAULT_WIDTH

Traits

DocAllocator

The DocAllocator trait abstracts over a type which can allocate (pointers to) Doc.

Print

PrintExt

PrintInCtx

PrintToString

PrintToStringInCtx

WriteColor

This trait describes the behavior of writers that support colored output.

Type Definitions

Alloc

Builder